requirements.txt
accelerate==0.29.3
datasets==2.19.0
diffusers==0.28.0.dev0
einops==0.8.0
Pillow==10.3.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.2
transformers==4.40.1
